Simple Example of Quest Marker adjustment
I made a modification to the dialogue system, so that I could demonstrate, within QuestTest, a system so that when one npc is talked to, the marker focuses on the other npc instead, and vice versa. At the moment the marker only appears when the target character is off screen... I could potentially create a different type of marker that appears when the npc is on screen as well. (Would need to make FadeImage specify which marker it wants to fade)