Modified Reward Manager to allow for creation of a UI element which stores all of the quests.
This was modeled after the version from the ShopManager - currently each object only involves a text element, but an image can easily be added if desired... This system will require manual implementation of the quests through the list I set up. I still need to implement a function for when buttons are clicked, as well as for functions to determine whether quests are completed depending on their quest type.